Refactor: MainPage 폴더 내의 폴더구조 변경
