+++ src/pages/DetailPage/components/DeleteModal.tsx
