+++ src/pages/DetailPage/index.tsx
@@ -6,8 +6,8 @@ import DetailHeader from './views/DetailHeader';
 import DetailBottom from './views/DetailBottom';
 import DetailPartySection from './views/DetailPartySection';
 import FixDetailHeader from './components/FixDetailHeader';
-import MatchApplyModal from '../MainPage/Components/MatchApply/MatchApplyModal';
-import MatchApplyButton from '../MainPage/Components/MatchApply/MatchApplyButton';
+import MatchApplyModal from '../MainPage/Components/MatchApplyButton/MatchApplyModal';
+import MatchApplyButton from '../MainPage/Components/MatchApplyButton/MatchApplyButton';
 import instance from '@apis';
 import ModalStore from '@stores/ModalStore';
 import { useMyPotIdStore } from '@stores/MyPotIdStore';
