+++ src/pages/MainPage/Components/SingleContent/SingleProfile.tsx
